# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "download_url": {
            "description": "URL autorizada de download da API (/attachments/{id}/download)",
            "title": "Download Url",
            "type": "string"
          },
          "entity_id": {
            "description": "UUID da entidade proprietária (site, structure, device, customer)",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "description": "Tipo da entidade vinculada",
            "title": "Entity Type",
            "type": "string"
          },
          "file_name": {
            "description": "Nome de arquivo seguro sanitizado",
            "title": "File Name",
            "type": "string"
          },
          "file_size_bytes": {
            "description": "Tamanho do arquivo em bytes",
            "minimum": 0.0,
            "title": "File Size Bytes",
            "type": "integer"
          },
          "id": {
            "description": "UUID do anexo armazenado",
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "description": "MIME type validado por inspeção de cabeçalho",
            "title": "Mime Type",
            "type": "string"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URL da miniatura otimizada se o arquivo for imagem suportada",
            "title": "Thumbnail Url"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "entity_id",
          "entity_type",
          "file_name",
          "mime_type",
          "file_size_bytes",
          "download_url",
          "version",
          "created_at"
        ],
        "title": "AttachmentRead",
        "type": "object"
      },
...
      "Body_api_v1_attachments_upload_attachment": {
        "properties": {
          "entity_id": {
            "description": "UUID da entidade associada (ex: estrutura, site, cliente)",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "description": "Tipo da entidade (structure, site, customer, device)",
            "title": "Entity Type",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Arquivo binário (JPEG, PNG, WebP ou PDF)",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "entity_id",
          "entity_type",
          "file"
        ],
        "title": "Body_api_v1_attachments_upload_attachment",
        "type": "object"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "description": "Senha atual do usuário",
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "description": "Nova senha com no mínimo 8 caracteres",
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "description": "E-mail corporativo do usuário",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Senha do usuário",
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "message": {
            "default": "Login realizado com sucesso",
            "description": "Mensagem de retorno",
            "title": "Message",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MeResponse",
            "description": "Dados do usuário logado"
          }
        },
        "required": [
          "user"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "email": {
            "description": "E-mail corporativo do usuário",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "description": "UUID do usuário autenticado",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "Nome de exibição do usuário",
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "description": "Lista de permissões granulares concedidas ao usuário",
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Função e perfil de acesso do usuário"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "permissions"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "UserRole": {
        "enum": [
          "admin",
          "engineer",
          "technician",
          "viewer"
        ],
        "title": "UserRole",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "FTTH Manager — Sistema de documentação física e óptica de rede FTTH",
    "title": "FTTH Manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/attachments": {
      "post": {
        "description": "Armazena arquivo de imagem ou PDF com verificação de tipo de conteúdo e isolamento de path traversal.",
        "operationId": "api_v1_attachments_upload_attachment",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_v1_attachments_upload_attachment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload de anexo ou foto de campo",
        "tags": [
          "Anexos e Fotos"
        ]
      }
    },
    "/api/v1/attachments/{attachment_id}": {
      "delete": {
        "description": "Remove anexo e seus metadados. Exige cabeçalho If-Match.",
        "operationId": "api_v1_attachments_attachment_id_delete_attachment",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "title": "Attachment Id",
              "type": "string"
            }
          },
          {
            "description": "Versão atual do recurso (If-Match)",
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "description": "Versão atual do recurso (If-Match)",
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Excluir anexo",
        "tags": [
          "Anexos e Fotos"
        ]
      }
    },
    "/api/v1/attachments/{attachment_id}/download": {
      "get": {
        "description": "Faz o download seguro de arquivo após validação das credenciais e permissões do usuário.",
        "operationId": "api_v1_attachments_attachment_id_download_download_attachment",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "title": "Attachment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download de anexo",
        "tags": [
          "Anexos e Fotos"
        ]
      }
    },
...
    "/api/v1/auth/change-password": {
      "post": {
        "description": "Atualiza a senha do usuário autenticado mediante confirmação da senha atual.",
        "operationId": "api_v1_auth_change_password_change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alterar senha do usuário atual",
        "tags": [
          "Autenticação"
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Autentica via e-mail e senha, define cookie HttpOnly seguro de sessão e rotaciona token.",
        "operationId": "api_v1_auth_login_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login de usuário",
        "tags": [
          "Autenticação"
        ]
      }
    },
...
  }
}